Atualização update userBy

+++ BackendBagaque/Services/Orders/OrdersService.cs
@@ -1,4 +1,5 @@
-﻿using BackendBagaque.Data;
+﻿using BackendBagaque.Models;
+using BackendBagaque.Data;
 
 namespace BackendBagaque.Services.Orders
 {
@@ -6,42 +7,48 @@ namespace BackendBagaque.Services.Orders
     {
         private readonly BagaqueDBContext context;
 
-        public OrdersService (BagaqueDBContext context)
+        public OrdersService(BagaqueDBContext context)
         {
             this.context = context;
         }
 
+        /*                  SERVICES ORDERS                     */ 
+        //Este END POINT Seleciona toda a tabela Order
         public List<Models.Orders> GetAll()
         {
             return context.Orders.ToList();
         }
 
+        //Este END POINT Seleciona por id a tabela Order
         public Models.Orders? GetOne(int IdOrders)
         {
             var orders = context.Orders.Find(IdOrders);
             return orders;
         }
 
-        public Models.Orders Create(Models.Orders orders)
+       //Este END POINT Cria dados para Tabela Order por id do Usuario ADm
+        public Models.Orders CreateOrdersByAdm(Models.Orders orders, int IdUsers)
         {
-            context.Orders.Add(orders);
-            context.SaveChanges();
-            return orders;
+            var user = context.Users.FirstOrDefault(u => u.IdUsers == IdUsers);
+            if (user == null || user.TypeUser != 2)
+            {
+                throw new Exception("Liberação só para usuario admin");
+            }
+            else
+            {
+                context.Orders.Add(orders);
+                context.SaveChanges();
+                return orders;
+            }
         }
 
-        public void Update(int IdOrders, Models.Orders orders)
+        //Este END POINT Atualiza os dados para Tabela Order por id do Usuario global
+        public void UpdateOrdersBy(int IdOrders, Models.Orders orders)
         {
             var ordersToUpdate = context.Orders.Find(IdOrders);
             if (ordersToUpdate != null)
             {
-                ordersToUpdate.Dater = orders.Dater;
-                ordersToUpdate.FinalDateDelivery = orders.FinalDateDelivery;
-                ordersToUpdate.CodeDelivery = orders.CodeDelivery;
-                ordersToUpdate.StatusOrder = orders.StatusOrder;
                 ordersToUpdate.TypePayment = orders.TypePayment;
-                ordersToUpdate.StatusPayment = orders.StatusPayment;
-                ordersToUpdate.IdUser = orders.IdUser;
-
                 context.SaveChanges();
             }
             else
@@ -50,18 +57,77 @@ namespace BackendBagaque.Services.Orders
             }
         }
 
-        public void Delete(int IdOrders)
+        //Este END POINT Atualiza os dados para Tabela Order por id do Usuario Adm
+        public void UpdateOrdersByAdmin(int IdOrders, Models.Orders orders, int IdUsers)
         {
-            var ordersToRemove = context.Orders.Find(IdOrders);
-            if (ordersToRemove != null)
+            var user = context.Users.FirstOrDefault(u => u.IdUsers == IdUsers);
+            if (user == null || user.TypeUser != 2)
             {
-                context.Orders.Remove(ordersToRemove);
-                context.SaveChanges();
+                throw new Exception("Liberação só para usuario admin");
             }
             else
             {
-                throw new Exception("Pedido não encontrado para o id " + IdOrders);
+                var ordersToUpdate = context.Orders.Find(IdOrders);
+                if (ordersToUpdate != null)
+                {
+                    ordersToUpdate.FinalDateDelivery = orders.FinalDateDelivery;
+                    ordersToUpdate.CodeDelivery = orders.CodeDelivery;
+                    ordersToUpdate.StatusOrder = orders.StatusOrder;
+                    ordersToUpdate.StatusPayment = orders.StatusPayment;
+                    ordersToUpdate.IdUser = orders.IdUser;
+
+                    context.SaveChanges();
+                }
+                else
+                {
+                    throw new Exception("Pedido não encontrado para o ID " + IdOrders);
+                }
             }
+
+        }
+
+        //Este END POINT Deleta os dados para Tabela Order por id do Usuario global
+        public void DeleteOrderByAdm(int IdOrders,int IdUsers)
+        {
+            var use = context.Users.FirstOrDefault(u => u.IdUsers == IdUsers);
+            if (use != null && use.TypeUser == 2)
+            {
+                var ordersToRemove = context.Orders.Find(IdOrders);
+                if (ordersToRemove != null)
+                {
+                    context.Orders.Remove(ordersToRemove);
+                    context.SaveChanges();
+                }
+                else
+                {
+                    throw new Exception("Pedido não encontrado para o id " + IdOrders);
+                }
+            }
+            else
+            {
+                throw new Exception("Liberação só para usuario admin");
+            }
+        }
+
+        /*                  SERVICES ORDERS/  PRODUCORDER          */
+        public List<Models.ProductOrder> GetProductOrderAll()
+        {
+            return context.ProductOrder.ToList();
+        }
+
+        //Este END POINT Seleciona por id a tabela Order
+        public Models.ProductOrder GetOneProductOrder(int IdProductOrder)
+        {
+            var productorder = context.ProductOrder.Find(IdProductOrder);
+            return productorder;
+        }
+
+        //Este END POINT Cria dados para Tabela ProductOrder por id do Usuario Global
+        public Models.ProductOrder CreateProductOrdersBy(Models.ProductOrder productorder)
+        {
+                context.ProductOrder.Add(productorder);
+                context.SaveChanges();
+                return productorder;
         }
     }
 }
